+++ CHANGELOG.rst
@@ -47,6 +47,13 @@ Changelog
   a single-valued RDN.
 * Added
   :func:`~cryptography.hazmat.primitives.asymmetric.ec.derive_private_key`.
+* Added support to
+  :meth:`~cryptography.hazmat.primitives.asymmetric.rsa.RSAPrivateKey.sign`
+  and
+  :meth:`~cryptography.hazmat.primitives.asymmetric.rsa.RSAPublicKey.verify`
+  RSA signatures with a
+  :class:`~cryptography.hazmat.primitives.asymmetric.utils.Prehashed`
+  digest.
 
 1.5.3 - 2016-11-05
 ~~~~~~~~~~~~~~~~~~